v3.4.0: connectivity refactor — utilities split, SampleContext adoption, dedup
Drives down the cross-module coupling identified by the v3.4.0 graphify
pass while keeping every existing `from utilities import ...` line in the
codebase working unchanged (utilities.py is now a re-export shim).

Helper consolidation:
- to_abs() collapsed from 4 copies (assemble_flye/hifiasm, html_reporter,
  preprocess_pacbio) into one defensive canonical version in bin/fs.py.
- validate_fasta() collapsed from 3 copies (curate_assembly,
  compare_assemblies, qc_assessment) into bin/fs.py, keeping the
  curate_assembly variant's empty-path guard.
- html_reporter's get_inat_obs stub: `except Exception` narrowed to
  `except ImportError` and the inline stubs given proper docstrings.

Module split (bin/utilities.py 750-line monolith -> 80-line shim):
- bin/sh.py:         run_subprocess_cmd, get_resource_values
- bin/log.py:        log_print, initialize_logging_environment,
                     generate_log_file (DEFAULT_LOG_FILE/ENVIRONMENT_TYPE
                     now declared at module load so log_print can't
                     NameError pre-init)
- bin/fs.py:         to_abs, validate_fasta, pigz_compress/decompress,
                     md5_check, move_file_up, sum_*_bases_with_pigz_safe,
                     calculate_genome_coverage
- bin/sample_csv.py: get_current_row_data, gen_sample_stats_dict,
                     analyze_nanostats, select_long_reads, plus the new
                     SampleContext dataclass, load_sample_context
                     helper, and AssemblerStage @runtime_checkable
                     Protocol.

SampleContext adoption (13 modules):
- assemble_flye/hifiasm/masurca/spades, preprocess_illumina/ont/pacbio/
  refseq, polish_assembly, curate_assembly, decontaminate_assembly/reads,
  and both qc_assessment call sites now replace the 6-line CSV-row
  preamble with `ctx = load_sample_context(...)`.
- Mixed `input_csvto_abs` / `input_csv_abs` / `output_dir_abs` /
  `output_dirto_abs` variants are gone; everything flows through
  ctx.input_csv / ctx.output_dir / ctx.current_series / ctx.current_row
  / ctx.current_index / ctx.sample_stats_dict.
- AssemblerStage annotations added wherever the (sample_id, input_csv,
  output_dir, cpu_threads, ram_gb) -> Optional[str] signature still
  matches; preprocess_illumina / decontaminate_reads / qc_assessment
  intentionally skipped (their return shape differs).

MaSuRCA Flye-vendor shims (bin/assemble_masurca.py):
- ensure_flye_vendor_shim    -> _masurca_vendor_flye_ensure_shim
- patch_environment_flye     -> _masurca_vendor_flye_patch_environment
- patch_flye_check           -> _masurca_vendor_flye_patch_check
- Orphan patch_check was missing wiring, not dead code: it applies the
  same PATH-Flye-acceptance fix to assemble.sh that patch_environment
  applies to environment.sh.  Wired in at line 577 *before* the
  skip_gap_closing_section call so the in-place patch isn't lost when
  the derivative assemble_skip_gap.sh is written.
- Fixed a latent shebang-splice bug in patch_check's defensive
  injection block (was reconstructing new_txt unchanged and appending
  inject at EOF instead of inserting after the shebang); the bug was
  dormant only because the function had no caller until now.

Stage: docstring annotations:
- Added Stage: blocks to 17 pipeline module docstrings (every
  preprocess_*, assemble_*, decontaminate_*, plus polish_, curate_,
  compare_, qc_, html_reporter, process_metadata, final_compress) so
  the next `/graphify --update` can wire implements_stage edges from
  the pipeline-flow SVG to the implementing code, eliminating the
  duplicate orphan stage communities.

Dockerfile:
- Added sh.py / log.py / fs.py / sample_csv.py to the wget list so the
  container build pulls in the new modules.

Net diff: ~199 insertions, ~864 deletions across 22 source files.
All 24 bin/*.py files pass py_compile cleanly.
